Working timer, popup and bar chart

+++ textprocessor.h
@@ -13,15 +13,15 @@ class TextProcessor : public QObject
 {
     Q_OBJECT
     Q_PROPERTY(QString text READ text WRITE setText NOTIFY textChanged)
-    Q_PROPERTY(int timer READ timer WRITE setTimer NOTIFY timerChanged)
+    Q_PROPERTY(qreal timer READ timer WRITE setTimer NOTIFY timerChanged)
 
 public:
     explicit TextProcessor(QObject *parent = 0);
 
     QString text() const;
     void setText(const QString& s);
-    int timer() const;
-    void setTimer(const int &t);
+    qreal timer() const;
+    void setTimer(const qreal& t);
 
     Q_INVOKABLE void randomiseText(const bool &);
     Q_INVOKABLE int checkState(const QString &);
@@ -40,7 +40,7 @@ private:
 
     QString m_text;
     QStringList textList;
-    int m_timer;
+    qreal m_timer;
     QTimer* clock;
     QFile file;
     bool typoFlag;
